formular: add 'number' field type
It checks that the bot is writing a number to the field. While this
is useful when we want to ensure that the data is correct for
simulations, it does not allow us to use the bots to put incorrect data
into the number field, such as a non-number string and test that wrong
input is caught. While one can get around this by not using `autofill`,
it may be a bug rather than a feature.