pjb-pilot: script to export data to csv
It is hard to parse json in R. This script extracts the most relevant
data and writes it to csv file.

Issues: There is a lot of annoying code to write to get this to work, a
lot of fidgety extraction. Automate repetitive parts so as not to have
to redo the same type of cleaning tasks.