config: max file size with error 413 page
nginx deals with identifying if a submitted file is too large and calls
the route for error 413. Implementing ourselves is not worth it, since
Firefox displays a 'connection reset' error if we do not read the whole
request. That is tricky code with little added value that nginx already
deals with properly. As long as the `client_max_body_size` in the nginx
config is smaller than the `#:max-form-data-file-length` in the
`(make-safety-limits ...)` call (see `dynamic.rkt`), nginx will throw
the error before the Racket server gets the file and display the error
page.

To avoid needlessly uploading large files, we will implement a client
side (JS) check to catch such uploads in the browser.