multi-review: admin stub for reviews
Pretty cool how easy it was to do this, although passing arguments to
the step being called looks like it will require `put`ting values and
then providing them with `provide-bindings` to a substudy for
modularity. Not too bad though, and it was easy to create the new
`button/dispatch` widget to allow dispatching to different steps.

TODO: the admin functionality and steps should be split out as an
admin-substudy rather than via named steps as I did for prototyping.