routes: redirect dashboard to identity server for identity users
A user who enrolled through an identity server should not see the
dashboard page on the congame (study) server. Instead they should be
redirected to see the studies available on the identity server.

Checking of user status is done through checking the start of the
display name. It may be preferable to do it by checking if
`identity-service-url` is set.