multi-review: use `parameterize` for admin area
This avoids all the ugly hacks passing around the shared data stored
with `put/instance` at the `(*root*)` stack, while the admin study has
to be run at the `(*root* admin)` substudy (or some other substudy).
Much cleaner, although the mechanism should not be abused and generally
speaking only makes sense for the admin and for shared data, even though
it is available with `put` as well.