multi-review: review reviews
review of reviews in admin area now allows to submit reviews of reviews
as well as edit them again and again. It also allows to
edit (first-order) reviews.

The review of the reviews is currently hardcoded towards research ideas,
so some refactoring will be needed to add this to pdf submissions.