multi-review,admin: button to remove reviews
This removes the review so it won't be seen by the user nor contribute
towards the score. It may be necessary to review this submission to
ensure that it has some score -- e.g. removing all reviews leads to an
undefined score. The removed reviews are stored in 'removed-reviews,
which may be helpful to provide future examples.