enroll: check enrollment code for identity users
We need to reparameterize the current-user since the current-user at
time of embedding is the api user. I do not understand why the `protect`
part leads to a failure -- presumably because the cookie is checked and
found to be that of the user when the embed is done, i.e. the api user.
But I don't quite grasp it, nor whether this will work later on or some
future redirect/get/forget/protect will fail.