IP notification: partial and full code generation, update planned features
Include only the first 40 lines of most code files in the IP
notification to keep it at around 100 pages instead of 260 pages. The
reduced version is used for the paper submission, the full version to
send along (in addition to the reduced) with the email.

I also updated the features. We should start making version updates with
features for better documentation (as most software does), as it is easy
to forget what we did.