dsl: deal also with step/study's and step's
We can refer to four types in transitions: handlers, studies, steps, and
step/studies. Confusingly, we create handlers with @step, so they still
need to be converted into steps. While this is hidden from users, it is
confusing.

Both handlers and studies have to be converted to steps, while
step/studies and steps (created directly with `make-step`, not via
@step) do not.