veneer: compile expressions inside @call
The current version of call does not allow special forms like `@get`
inside of `@call`. Thus the following failed:

`@call[add1 @get[counter]]`

since the special form of `get` turns this into `(get 'counter)`, while
the form as read by the scribble reader yields `(get counter)`.

There is a question whether we should have the special form of get do
this extra work or not, since it makes it syntactically inconsistent
with full racket. But even if we change that, we should compile
expressions inside `@call`.