bug: create study to trigger the bug
We are using up exponentially more memory and time as we keep repeating
a substate. Note that for some reason it is crucial that we do those
initial steps that are part of another substudy. It may well be that we
don't need to have multiple of them. I need between 5 and 8 repetitions
before the lag becomes noticeable. As the study becomes smaller, the
number of repetitions is probably going up.