studies,edpb: study to upload abstracts as csv
This study allows uploading a csv file for abstracts within a study. It
is meant to be used as admin of a study instance prior to launching it
to users.

One decision to double check is whether to store the abstracts in memory
via *ABSTRACTS* or whether to have it available only via
`get/instance/top 'abstracts`. The former will not yet work
as *ABSTRACTS* is not yet provided, but should require fewer DB calls.