wip: split study into intro and main
Intro exists to elicit the patience level and tell them how to signup.
The patience level is elicited here to see if we lose impatient people
due to requiring signup. The signup is needed so that people can resume
the study on a later day, since the study requires multiple days of
participation.