edpb,admin: list participants to be approved
Next step (when I need it): order them by when they enrolled, so that I
can distinguish those that I (presumably) approved in the past from
those that I need to approve now. A better way is to have a new batch
number that I change prior to each new run of prolific participants, so
that I could have multiple runs on the same day, or simply better tracking.