edpb pilot: update for more tasks and 66% success
Participants were done in roughly 15 minutes with the main study, and
only took 12 seconds per abstract, so we need to ask for more. We
require 66% success on abstracts to avoid the 20-33% of people who just
click randomly and rush through the study. This requires highlighting
how many tasks they got wrong so far.

Probably we'll need to make the task harder, even with many tasks it is
too easy.