conscript/local: temporary fix for form bug
This fix is temporary, because it seems wrong to use as the request the
request of the original study. So far, this didn't cause any errors, but
I feel that this is not the right approach.

Bug in previous version: in `preview`, if a form got submitted on page 1
and on page 2 we have another form, then somehow the request persisted
in such a way that the form on page 2 was processed with the values from
the previous form. If both forms required a number, then the first form
would be submitted successfully, and then second form would two,
literally skipping the page!