docs: add inline note
Scribble's `margin-note` can be awkward because at mobile sizes (i.e. when displayed inside the main column) the note seems to relate to the preceding paragraph, but at desktop sizes it gets placed in the margin alongside the _following_ paragraph. This alternative forces the note inline for more control, also allows use of an icon with #:type.